+++ lib/concurrent-ruby/concurrent/executor/cached_thread_pool.rb
@@ -51,10 +51,9 @@ module Concurrent
     def ns_initialize(opts)
       super(opts)
       if Concurrent.on_jruby?
-        self.auto_terminate = opts.fetch(:auto_terminate, true)
         @max_queue          = 0
         @executor           = java.util.concurrent.Executors.newCachedThreadPool(
-            DaemonThreadFactory.new(self.auto_terminate?))
+            DaemonThreadFactory.new(ns_auto_terminate?))
         @executor.setRejectedExecutionHandler(FALLBACK_POLICY_CLASSES[@fallback_policy].new)
         @executor.setKeepAliveTime(opts.fetch(:idletime, DEFAULT_THREAD_IDLETIMEOUT), java.util.concurrent.TimeUnit::SECONDS)
       end