+++ lib/concurrent-ruby/concurrent/executor/java_thread_pool_executor.rb
@@ -108,35 +108,17 @@ if Concurrent.on_jruby?
           queue = java.util.concurrent.LinkedBlockingQueue.new(@max_queue)
         end
 
-        self.auto_terminate = opts.fetch(:auto_terminate, true)
-
         @executor = java.util.concurrent.ThreadPoolExecutor.new(
             min_length,
             max_length,
             idletime,
             java.util.concurrent.TimeUnit::SECONDS,
             queue,
-            DaemonThreadFactory.new(self.auto_terminate?),
+            DaemonThreadFactory.new(ns_auto_terminate?),
             FALLBACK_POLICY_CLASSES[@fallback_policy].new)
 
       end
     end
 
-    class DaemonThreadFactory
-      # hide include from YARD
-      send :include, java.util.concurrent.ThreadFactory
-
-      def initialize(daemonize = true)
-        @daemonize = daemonize
-      end
-
-      def newThread(runnable)
-        thread = java.util.concurrent.Executors.defaultThreadFactory().newThread(runnable)
-        thread.setDaemon(@daemonize)
-        return thread
-      end
-    end
-
-    private_constant :DaemonThreadFactory
   end
 end