+++ spec/concurrent/executor/executor_quits.rb
@@ -4,11 +4,11 @@ $LOAD_PATH.push lib unless $LOAD_PATH.include? lib
 
 require 'concurrent-ruby'
 
-# the test relies on replicating that Minitest messed up the AtExit handling
-Concurrent.disable_at_exit_handlers!
-pool = Concurrent::CachedThreadPool.new
-pool.post do
-  sleep # sleep indefinitely
+executors = [Concurrent::CachedThreadPool.new, Concurrent::SingleThreadExecutor.new, Concurrent::FixedThreadPool.new(1)]
+executors.each do |executor|
+  executor.post do
+    sleep # sleep indefinitely
+  end
 end
 
 # the process main thread should quit out which should kill the daemon CachedThreadPool